changes according to review comments
